# tests/test_dashboard/snapshots/openapi.json
{
  "components": {
    "schemas": {
      "ArtifactDirectoryFile": {
        "additionalProperties": true,
        "properties": {
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "name",
          "path",
          "icon"
        ],
        "title": "ArtifactDirectoryFile",
        "type": "object"
      },
      "ArtifactDirectoryResponse": {
        "additionalProperties": true,
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/ArtifactDirectoryFile"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "ArtifactDirectoryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResearchArtifact": {
        "additionalProperties": true,
        "properties": {
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "name",
          "path",
          "icon"
        ],
        "title": "ResearchArtifact",
        "type": "object"
      },
      "ResearchResponse": {
        "additionalProperties": true,
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ResearchArtifact"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "main_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main File"
          }
        },
        "required": [
          "artifacts"
        ],
        "title": "ResearchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP surface for the local Spec Kitty dashboard. All routes are localhost-only. See `kitty-specs/frontend-api-fastapi-openapi-migration-01KQN2JA/` for the contract authoring rules.",
    "title": "Spec Kitty Dashboard API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/artifact/{feature_id}/{name}": {
      "get": {
        "operationId": "get_artifact_api_artifact__feature_id___name__get",
        "parameters": [
          {
            "in": "path",
            "name": "feature_id",
            "required": true,
            "schema": {
              "title": "Feature Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/api/checklists/{feature_id}": {
      "get": {
        "operationId": "list_checklists_api_checklists__feature_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "feature_id",
            "required": true,
            "schema": {
              "title": "Feature Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactDirectoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Checklists",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/checklists/{feature_id}/{file_name}": {
      "get": {
        "operationId": "get_checklist_file_api_checklists__feature_id___file_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "feature_id",
            "required": true,
            "schema": {
              "title": "Feature Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "file_name",
            "required": true,
            "schema": {
              "title": "File Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Checklist File",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/contracts/{feature_id}": {
      "get": {
        "operationId": "list_contracts_api_contracts__feature_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "feature_id",
            "required": true,
            "schema": {
              "title": "Feature Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactDirectoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Contracts",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/contracts/{feature_id}/{file_name}": {
      "get": {
        "operationId": "get_contract_file_api_contracts__feature_id___file_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "feature_id",
            "required": true,
            "schema": {
              "title": "Feature Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "file_name",
            "required": true,
            "schema": {
              "title": "File Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Contract File",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/api/dossier/artifacts": {
      "get": {
        "operationId": "list_artifacts_api_dossier_artifacts_get",
        "parameters": [
          {
            "in": "query",
            "name": "feature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feature"
            }
          },
          {
            "in": "query",
            "name": "class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class"
            }
          },
          {
            "in": "query",
            "name": "wp_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Wp Id"
            }
          },
          {
            "in": "query",
            "name": "step_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Step Id"
            }
          },
          {
            "in": "query",
            "name": "required_only",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Required Only"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Artifacts",
        "tags": [
          "dossier"
        ]
      }
    },
    "/api/dossier/artifacts/{artifact_key}": {
      "get": {
        "operationId": "get_artifact_detail_api_dossier_artifacts__artifact_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_key",
            "required": true,
            "schema": {
              "title": "Artifact Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "feature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact Detail",
        "tags": [
          "dossier"
        ]
      }
    },
    "/api/dossier/overview": {
      "get": {
        "operationId": "get_overview_api_dossier_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "feature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Overview",
        "tags": [
          "dossier"
        ]
      }
    },
    "/api/dossier/snapshots/export": {
      "get": {
        "operationId": "export_snapshot_api_dossier_snapshots_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "feature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Snapshot",
        "tags": [
          "dossier"
        ]
      }
    },
...
    "/api/research/{feature_id}": {
      "get": {
        "operationId": "get_research_api_research__feature_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "feature_id",
            "required": true,
            "schema": {
              "title": "Feature Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Research",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/research/{feature_id}/{file_name}": {
      "get": {
        "operationId": "get_research_file_api_research__feature_id___file_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "feature_id",
            "required": true,
            "schema": {
              "title": "Feature Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "file_name",
            "required": true,
            "schema": {
              "title": "File Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Research File",
        "tags": [
          "artifacts"
        ]
      }
    },
...
  }
}